gameOver function added
Added a gameover function that checks if any of the blocks in my block array contains goal tile. for every block that contains goal tile add 1 new block to my finished block array.

Ones the finished block array is as big as the block array the game will end and the board will be removed.

2 do:

Make function to move the boxes!